feat(storybook-config): add scss config to preset webpack
