+++ packages/react/.storybook/webpack.config.js
@@ -6,23 +6,16 @@
  */
 
 const path = require('path');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
-const rtlcss = require('rtlcss');
-const customProperties = require('postcss-custom-properties');
 
 const {
-  CARBON_REACT_STORYBOOK_USE_CUSTOM_PROPERTIES = 'false',
   CARBON_REACT_STORYBOOK_USE_EXTERNAL_CSS,
   CARBON_REACT_USE_CONTROLLED_STATE_WITH_EVENT_LISTENER,
-  CARBON_REACT_STORYBOOK_USE_RTL,
   NODE_ENV = 'development',
 } = process.env;
 
-const useExternalCss = NODE_ENV === 'production';
 const useControlledStateWithEventListener =
   CARBON_REACT_USE_CONTROLLED_STATE_WITH_EVENT_LISTENER === 'true';
-const useRtl = CARBON_REACT_STORYBOOK_USE_RTL === 'true';
 
 const replaceTable = {
   useControlledStateWithEventListener,
@@ -93,78 +86,6 @@ module.exports = ({ config, mode }) => {
     ],
     enforce: 'pre',
   });
-
-  const sassLoader = {
-    loader: 'sass-loader',
-    options: {
-      prependData() {
-        return `
-          $feature-flags: (
-            ui-shell: true,
-            enable-css-custom-properties: ${CARBON_REACT_STORYBOOK_USE_CUSTOM_PROPERTIES},
-          );
-        `;
-      },
-      sassOptions: {
-        includePaths: [path.resolve(__dirname, '..', 'node_modules')],
-      },
-      sourceMap: true,
-    },
-  };
-
-  const fastSassLoader = {
-    loader: 'fast-sass-loader',
-    options: {
-      data: `
-        $feature-flags: (
-          ui-shell: true,
-          enable-css-custom-properties: ${CARBON_REACT_STORYBOOK_USE_CUSTOM_PROPERTIES},
-        );
-      `,
-    },
-  };
-
-  config.module.rules.push({
-    test: /\.scss$/,
-    sideEffects: true,
-    use: [
-      {
-        loader: useExternalCss ? MiniCssExtractPlugin.loader : 'style-loader',
-      },
-      {
-        loader: 'css-loader',
-        options: {
-          importLoaders: 2,
-          sourceMap: true,
-        },
-      },
-      {
-        loader: 'postcss-loader',
-        options: {
-          plugins: () => {
-            const autoPrefixer = require('autoprefixer')({
-              overrideBrowserslist: ['last 1 version', 'ie >= 11'],
-            });
-            return [
-              customProperties(),
-              autoPrefixer,
-              ...(useRtl ? [rtlcss] : []),
-            ];
-          },
-          sourceMap: true,
-        },
-      },
-      NODE_ENV === 'production' ? sassLoader : fastSassLoader,
-    ],
-  });
-
-  if (useExternalCss) {
-    config.plugins.push(
-      new MiniCssExtractPlugin({
-        filename: '[name].[contenthash].css',
-      })
-    );
-  }
 
   config.resolve = {
     modules: ['node_modules'],
